Refactor query test cases into subfolder
